Use the same struct for keeping roots as what is passed into Tracy.
This allows to avoid some duplication that would otherwise be needed by
having a separate "julia version" of the source location with the actual
julia string sin it.

This also makes this struct mutable since it simplifies updating it. We
need the source location to live on the heap anyway since we are passing
a pointer to it to Tracy.