Connect to a server during testing to ensure tracy likes the data we emit.
In this commit, we spawn a headless tracy server and connect it to a
process that runs instrumented code. This ensures that the checks in the
tracy server are being run which provides extra verification that our
implementation is correct.

In addition to this, we are now emitting a CSV file from the recorded profile data. This allows us to verify that the zones are being sent to Tracy as intended.

Finally, we've made a quality-of-life improvement to `runtests.jl`. By setting `connect_tracy_gui` to `true`, we automatically connected to a process running the instrumented thereby facilitating the manual inspection of profiling results.