Add `wait_for_tracy()` export
This gives user code that is running on a non-instrumented build a way
to replicate the logic from `JULIA_WAIT_FOR_TRACY=1`, e.g. waiting until
the tracy profiler connects.  This allows for easier synchronization of
relevant workload, e.g. starting a task that will run the workload which
immediately blocks on waiting for tracy, then starting up the capture
agent.