Add Preferences.jl support
This change adds preliminary Preferences.jl support, and also removes
the dispatch-based "compile-time" tracepoint configuration we had
before.

A `Preferences.toml` might look like:
```
[MyPkg.Tracy]
enabled = true
whitelist = ['foo', 'bar.*']
blacklist = ['baz']
```

The loading behavior is somewhat non-standard, since it allows for
preferences to be inherited from the `Preferences.toml` of the
instrumented package.

`Preferences.toml` can be copied to/from `LocalPreferences.toml`. They
have exactly the same format, and the keys are chosen to be easy to
remember and edit by hand.

Preferences is not my favorite package, but it has superpowers (mostly
to trigger pre-compilation from top-level configuration/data) that we
can't easily do without so this seems like the right way to go for now.