Use `Base.Linking.private_libdir()` for base libTracyClient
It is a problem for relocatability for us to resolve the path
dynamically from `dllist()` at pre-compile time, so instead we resolve
it at `__init__()` by looking directly in the private libdir.

Eventually, this should be moved into its own stdlib, in which case we
can use the library from LibTracyClient_jll directly.